+++ ui/src/com/kotcrab/vis/ui/widget/VisValidatableTextField.java
@@ -88,6 +88,7 @@ public class VisValidatableTextField extends VisTextField {
 
 	private void init () {
 		setProgrammaticChangeEvents(true);
+		setIgnoreEqualsTextChange(false);
 	}
 
 	@Override